(PC-31826)[PRO] style: Reorganise styles/components.
