+++ pro/src/pages/LostPassword/LostPassword.tsx
@@ -8,7 +8,7 @@ import { useInitReCaptcha } from 'hooks/useInitReCaptcha'
 import { useNotification } from 'hooks/useNotification'
 import { useRedirectLoggedUser } from 'hooks/useRedirectLoggedUser'
 import logoPassCultureProFullIcon from 'icons/logo-pass-culture-pro-full.svg'
-import logoStyles from 'styles/components/layout/_Logo.module.scss'
+import logoStyles from 'styles/components/_Logo.module.scss'
 import { Hero } from 'ui-kit/Hero/Hero'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
 import { getReCaptchaToken } from 'utils/recaptcha'
