+++ pro/src/pages/ResetPassword/ResetPassword.tsx
@@ -6,7 +6,7 @@ import { api } from 'apiClient/api'
 import { AppLayout } from 'app/AppLayout'
 import { useRedirectLoggedUser } from 'hooks/useRedirectLoggedUser'
 import logoPassCultureProFullIcon from 'icons/logo-pass-culture-pro-full.svg'
-import logoStyles from 'styles/components/layout/_Logo.module.scss'
+import logoStyles from 'styles/components/_Logo.module.scss'
 import { Hero } from 'ui-kit/Hero/Hero'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
 import { parse } from 'utils/query-string'
