+++ pro/src/styles/components/_CookieModals.scss
