+++ pro/src/styles/components/_Logo.module.scss
