+++ pro/src/styles/components/index.scss
@@ -1,1 +1,1 @@
-@import "layout/index";
+@import "CookieModals";
