--- pro/src/styles/components/layout/index.scss
@@ -1,1 +1,0 @@
-@import "CookieModals";
